#110 - migrate server layer to @vulps22/dynamic-endpoint-router
