+++ src/routes/api/v1/vote.ts
@@ -1,8 +1,9 @@
 import crypto from 'crypto';
-import express, { Request, Response } from 'express';
-import { Logger } from '../utils';
-import { inventoryService, userService } from '../services';
-import { Storable } from '../types';
+import { Request, Response, NextFunction } from 'express';
+import { ApiRoute, rawBody } from '@vulps22/dynamic-endpoint-router';
+import { Logger } from '../../../utils';
+import { inventoryService, userService } from '../../../services';
+import { Storable } from '../../../types';
 
 interface TopGGUser {
     id: string;
@@ -30,10 +31,23 @@ interface TopGGVotePayload {
     };
 }
 
-function verifySignature(secret: string, rawBody: Buffer, signatureHeader: string): string | null {
-    // Header format: "t=<timestamp>,v1=<hmac>"
+function verifyTopgg(req: Request, res: Response, next: NextFunction): void {
+    const header = req.headers['x-topgg-signature'];
+
+    if (!header || typeof header !== 'string') {
+        Logger.error('[VoteWebhook] Missing signature header');
+        res.status(401).send('Unauthorized');
+        return;
+    }
+
+    const secret = process.env.TOPGG_WEBHOOK_AUTH;
+    if (!secret) {
+        res.status(503).send('Service Unavailable');
+        return;
+    }
+
     const parts = Object.fromEntries(
-        signatureHeader.split(',').map(part => {
+        header.split(',').map(part => {
             const idx = part.indexOf('=');
             return [part.slice(0, idx), part.slice(idx + 1)] as [string, string];
         })
@@ -41,54 +55,41 @@ function verifySignature(secret: string, rawBody: Buffer, signatureHeader: strin
     const timestamp = parts['t'];
     const theirSignature = parts['v1'];
 
-    if (!timestamp) return 'missing timestamp (t) in signature header';
-    if (!theirSignature) return 'missing signature (v1) in signature header';
+    if (!timestamp || !theirSignature) {
+        Logger.error('[VoteWebhook] Invalid signature header format');
+        res.status(401).send('Unauthorized');
+        return;
+    }
 
-    const message = `${timestamp}.${rawBody.toString('utf8')}`;
+    const message = `${timestamp}.${(req.body as Buffer).toString('utf8')}`;
     const ourSignature = crypto.createHmac('sha256', secret).update(message).digest('hex');
 
     const ours = Buffer.from(ourSignature);
     const theirs = Buffer.from(theirSignature);
 
-    if (ours.length !== theirs.length) return `signature length mismatch (got ${theirs.length} bytes, expected ${ours.length})`;
+    if (ours.length !== theirs.length || !crypto.timingSafeEqual(ours, theirs)) {
+        Logger.error('[VoteWebhook] Invalid signature');
+        res.status(401).send('Unauthorized');
+        return;
+    }
 
-    if (!crypto.timingSafeEqual(ours, theirs)) return 'HMAC mismatch — secret may be wrong or body was modified in transit';
-
-    return null; // null = valid
+    next();
 }
 
-export function createVoteWebhookApp(webhookSecret: string): express.Application {
-    const app = express();
-
-    app.use(express.raw({ type: 'application/json' }));
-
-    app.post('/webhook/vote', async (req: Request, res: Response): Promise<void> => {
-        const signatureHeader = req.headers['x-topgg-signature'];
-
-        if (!signatureHeader || typeof signatureHeader !== 'string') {
-            Logger.error(`[VoteWebhook] Missing signature header`);
-            res.status(401).send('Unauthorized');
-            return;
-        }
-
-        const signatureError = verifySignature(webhookSecret, req.body as Buffer, signatureHeader);
-        if (signatureError !== null) {
-            Logger.error(`[VoteWebhook] Invalid signature — ${signatureError} (header: ${signatureHeader.substring(0, 60)}...)`);
-            res.status(401).send('Unauthorized');
-            return;
-        }
-
+export const route: ApiRoute = {
+    middleware: [rawBody(), verifyTopgg],
+    async post(req: Request, res: Response): Promise<void> {
         let payload: TopGGVotePayload;
         try {
             payload = JSON.parse((req.body as Buffer).toString('utf8')) as TopGGVotePayload;
         } catch {
-            Logger.error(`[VoteWebhook] Failed to parse request body`);
+            Logger.error('[VoteWebhook] Failed to parse request body');
             res.status(400).send('Bad Request');
             return;
         }
 
         if (payload.type === 'webhook.test') {
-            Logger.log(`[VoteWebhook] Test webhook received, ignoring.`);
+            Logger.log('[VoteWebhook] Test webhook received, ignoring.');
             res.status(200).send('OK');
             return;
         }
@@ -120,15 +121,5 @@ export function createVoteWebhookApp(webhookSecret: string): express.Application
         Logger.log(`[VoteWebhook] Vote received — user: ${discordUserId}, weight: ${payload.data.weight ?? 1}`);
 
         res.status(200).send('OK');
-    });
-
-    return app;
-}
-
-export function startVoteWebhook(port: number, webhookSecret: string): void {
-    const app = createVoteWebhookApp(webhookSecret);
-
-    app.listen(port, () => {
-        Logger.log(`[VoteWebhook] Listening on port ${port}`);
-    });
-}
+    },
+};
